remove discoverhue (python3 lib), switch to ssdp.py by DanKrause
