--- script.service.hue/resources/lib/discoverhue/__init__.py
@@ -1,2 +1,0 @@
-""" Auto discovery of Hue bridges """
-from .discoverhue import find_bridges
